# backend/api_docs.yaml
swagger: '2.0'
info:
  version: "1"
  title: "MusiCloud API"
  description: >-
    This is the REST API documentation for the MusiCloud music mixing platform.
host: "dcumusicloud.com"
basePath: "/api/v1"
...
schemes:
  - "http"
paths:
...
  /auth/verify:
    get:
      tags:
      - "auth"
      summary: "Verify new user's email"
      description: "When the user clicks the link sent to their email, there account will be verified."
      produces:
      - "application/json"
      parameters:
        - in: "query"
          name: "code"
          required: true
          type: "string"
          description: "The randomly generated code emailed to the user on account creation."
      responses:
        200:
          description: "successful query"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Verified."
        400:
          description: "Bad info sent, or info missing."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Invalid code."
        503:
          description: "Server is unavailable"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "MySQL unavailable"
  /auth/login:
    post:
      tags:
      - "auth"
      summary: "Login using a username and a password"
      description: "By passing a user's username along with their password to this request as part of the request body, if the user exists in the database, an access token will be returned to them. The user should be verified in order to log in."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User to be logged in."
        required: true
        schema:
          type: "object"
          properties:
            username:
              type: "string"
              example: "Herman"
            password:
              type: "string"
              example: "SecurePassword"
      responses:
        200:
          description: "Logged in"
          schema:
            type: "object"
            properties:
              access_token:
                type: "string"
                example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6MywidXNlcm5hbWUiOiJFZGRpZSBNdXJwaHkiLCJzdHVkZW50X2lkIjoiMTQ1Nzc4MDAiLCJmaXJzdF9uYW1lIjoiRW1pbHkiLCJzdXJuYW1lIjoiTWNHaXZlcm4iLCJ2ZXJpZmllZCI6MSwiZW1haWwiOiJlbWlseS5tY2dpdmVybjRAbWFpbC5kY3UuaWUiLCJwYXNzd3JkIjoiJDJhJDEwJERJSElubVBTWVJvNjhqUUd6TC9HVWU2cUxQaWtNSVY2VGtyc29PTzhaTXU4Z2hSZlJBTE95In0.fPZVIxg5a77bm08ErdWtxgYw3B_IgfNxO8BAsWAimyo"
        401:
          description: "Bad info sent, or info missing."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Bad login credentials."
        403:
          description: "The account has yet to be verified."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Account not verified."
        422:
          description: "Malformed request body."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "A req validation traceback will be here."
        503:
          description: "Server is unavailable"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "MySQL unavailable"
  /auth/logout:
    post:
      tags:
      - "auth"
      summary: "Deauthorise a token"
      description: "By passing an access_token to this endpoint in the request header, the corrisponding token will be removed the Logins table. This means that the passed token will no longer be valid for use with access_token protected endpoints."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Header"
          description: "This is where you should set your access_token."
          required: true
          type: "string"
      responses:
        200:
          description: "successful query"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "User has been successfully logged out!"
        422:
          description: "Malformed request body."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "A req validation traceback will be here."
        503:
          description: "Server is unavailable"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "MySQL unavailable."
...